refact: 코트 테스트 후 수정 1차
1. dto request/response 패키지 구분
2. s3 > application.properties로 통일
3. 좋아요 조회 -> 보드id 없을 때 not found 수정
4. 리뷰등록 -> 게임을 찾을 수 없음으로 수정
5. 게임추천 조회-> 경로에서 userId 삭제\
6. 파일 확장자(이미지) 체크 로직, 사이즈 크기 최대 5mb > 크키는 추후 수정 예정
7. controller.service 주석 작성
