+++ src/main/java/com/example/gamemate/domain/game/controller/GameRecommendContorller.java
@@ -1,8 +1,8 @@
 package com.example.gamemate.domain.game.controller;
 
-import com.example.gamemate.domain.game.dto.GameRecommendHistorysResponseDto;
-import com.example.gamemate.domain.game.dto.UserGamePreferenceRequestDto;
-import com.example.gamemate.domain.game.dto.UserGamePreferenceResponseDto;
+import com.example.gamemate.domain.game.dto.response.GameRecommendHistorysResponseDto;
+import com.example.gamemate.domain.game.dto.request.UserGamePreferenceRequestDto;
+import com.example.gamemate.domain.game.dto.response.UserGamePreferenceResponseDto;
 import com.example.gamemate.domain.game.service.GameRecommendService;
 import com.example.gamemate.global.config.auth.CustomUserDetails;
 import lombok.RequiredArgsConstructor;
@@ -44,17 +44,15 @@ public class GameRecommendContorller {
     /**
      * 특정 사용자의 게임 추천 이력을 조회합니다.
      *
-     * @param userId 조회할 사용자의 ID
      * @param customUserDetails 인증된 사용자 정보
      * @return 게임 추천 이력 목록을 포함한 ResponseEntity
      */
-    @GetMapping("/{userId}")
+    @GetMapping
     public ResponseEntity<Page<GameRecommendHistorysResponseDto>> getGameRecommendHistories(
-            @PathVariable Long userId,
             @AuthenticationPrincipal CustomUserDetails customUserDetails
     ) {
 
-        Page<GameRecommendHistorysResponseDto> responseDto = gameRecommendService.getGameRecommendHistories(userId, customUserDetails.getUser());
+        Page<GameRecommendHistorysResponseDto> responseDto = gameRecommendService.getGameRecommendHistories(customUserDetails.getUser());
         return new ResponseEntity<>(responseDto, HttpStatus.OK);
     }
