+++ src/main/java/com/example/gamemate/domain/game/dto/request/GameEnrollRequestCreateRequestDto.java
@@ -1,21 +1,21 @@
-package com.example.gamemate.domain.game.dto;
+package com.example.gamemate.domain.game.dto.request;
 
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 @Getter
-@NoArgsConstructor
-public class GameUpdateRequestDto {
+
+public class GameEnrollRequestCreateRequestDto {
     private String title;
     private String genre;
     private String platform;
     private String description;
 
 
-    public GameUpdateRequestDto(String title, String genre, String platform, String description) {
+    public GameEnrollRequestCreateRequestDto(String title, String genre, String platform, String description) {
         this.title = title;
         this.genre = genre;
         this.platform = platform;
         this.description = description;
+
     }
 }
