+++ src/main/java/com/example/gamemate/domain/game/dto/response/ChatResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.gamemate.domain.game.dto;
+package com.example.gamemate.domain.game.dto.response;
 
 import lombok.*;
