+++ src/main/java/com/example/gamemate/domain/game/dto/response/GameFindAllResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.gamemate.domain.game.dto;
+package com.example.gamemate.domain.game.dto.response;
 
 import com.example.gamemate.domain.game.entity.Game;
 import com.example.gamemate.domain.review.entity.Review;
