+++ src/main/java/com/example/gamemate/domain/game/service/GameRecommendService.java
@@ -1,6 +1,9 @@
 package com.example.gamemate.domain.game.service;
 
-import com.example.gamemate.domain.game.dto.*;
+import com.example.gamemate.domain.game.dto.response.GameRecommendHistorysResponseDto;
+import com.example.gamemate.domain.game.dto.response.GameRecommendationResponseDto;
+import com.example.gamemate.domain.game.dto.request.UserGamePreferenceRequestDto;
+import com.example.gamemate.domain.game.dto.response.UserGamePreferenceResponseDto;
 import com.example.gamemate.domain.game.entity.GameRecommendHistory;
 import com.example.gamemate.domain.game.entity.UserGamePreference;
 import com.example.gamemate.domain.game.repository.GameRecommendHistoryRepository;
@@ -31,6 +34,7 @@ public class GameRecommendService {
     private final GameRecommendHistoryRepository gameRecommendHistoryRepository;
     private final GeminiService geminiService;
 
+    // 게임 추천 요청 및 응답
     @Transactional
     public UserGamePreferenceResponseDto createUserGamePreference(UserGamePreferenceRequestDto requestDto, User loginUser) {
 
@@ -49,14 +53,18 @@ public class GameRecommendService {
 
         UserGamePreference saveData = userGamePreferenceRepository.save(userGamePreference);
 
-        // 저장된 선호도를 기반으로 Gemini API 호출
         String prompt = String.format(
                 "나에게 맞는 게임 3개 추천해줘 선호하는 장르는 %s이고 플레이 스타일은 %s 정도고 플레이 타임은 %s 정도고 난이도는 %s 그리고 플랫폼은 %s이고 추가적인 요청은 %s 야 " +
-                        "응답은 제목(title), 간단한 내용(description), 평점(star), 나와의 매칭점수(matchingScore), 추천 이유(reasonForRecommendation)를 적어주고 " +
-                        "응답은 순수 JSON 배열로 알려",
+                        "응답은 " +
+                        "한글(영어)로 된 제목(title), " +
+                        "간단한 내용(description)," +
+                        "metacriticScore 점수(metacriticScore)," +
+                        "나와의 매칭점수(matchingScore)," +
+                        "추천 이유(reasonForRecommendation)를 적어주고 " +
+                        "응답은 순수 JSON 배열로 알려줘",
                 userGamePreference.getPreferredGenres(),
-                userGamePreference.getPlayTime(),
                 userGamePreference.getPlayStyle(),
+                userGamePreference.getPlayTime(),
                 userGamePreference.getDifficulty(),
                 userGamePreference.getPlatform(),
                 userGamePreference.getExtraRequest()
@@ -64,12 +72,12 @@ public class GameRecommendService {
 
         String recommendation = geminiService.getContents(prompt);
         log.info(recommendation);
-        // 추천 JSON 문자열을 구조화된 객체로 파싱
+
 
         ObjectMapper objectMapper = new ObjectMapper();
         List<GameRecommendationResponseDto> gameRecommendations;
         try {
-            // 응답 문자열에서 JSON 배열 추출
+
             String jsonArray = recommendation
                     .replaceAll("(?s)^.*?```json\\s*", "") // 시작 부분의 ```
                     .replaceAll("\\s*```\\s*$", "") // 끝 부분의 ```
@@ -81,7 +89,6 @@ public class GameRecommendService {
             throw new ApiException(ErrorCode.RECOMMENDATION_NOT_FOUND);
         }
 
-        // GameRecommendHistory 엔티티 생성 및 저장
         List<GameRecommendHistory> gameRecommendHistories = new ArrayList<>();
         for (GameRecommendationResponseDto responseDto : gameRecommendations) {
             GameRecommendHistory history = new GameRecommendHistory(
@@ -90,27 +97,23 @@ public class GameRecommendService {
                     responseDto.getDescription(),
                     responseDto.getMatchingScore(),
                     responseDto.getReasonForRecommendation(),
-                    responseDto.getStar(),
+                    responseDto.getMetacriticScore(),
                     saveData
             );
             gameRecommendHistories.add(history);
         }
-        // 저장
+
         gameRecommendHistoryRepository.saveAll(gameRecommendHistories);
 
         return new UserGamePreferenceResponseDto(saveData, gameRecommendations);
     }
 
-    public Page<GameRecommendHistorysResponseDto> getGameRecommendHistories(Long userId, User loginUser) {
-
-        if (userId != loginUser.getId()){
-            throw new ApiException(ErrorCode.FORBIDDEN);
-        }
+    //게임 추천 조회
+    public Page<GameRecommendHistorysResponseDto> getGameRecommendHistories( User loginUser) {
 
         Pageable pageable = PageRequest.of(0, 15);
-        Page<GameRecommendHistory> histories = gameRecommendHistoryRepository.findByUserId(userId, pageable);
+        Page<GameRecommendHistory> histories = gameRecommendHistoryRepository.findByUserId(loginUser.getId(), pageable);
 
-       // return histories.map(gameRecommendHistory -> new GameRecommendHistorysResponseDto(gameRecommendHistory));
         return histories.map(GameRecommendHistorysResponseDto::new);
 
     }
