+++ src/main/java/com/example/gamemate/domain/like/dto/response/ReviewLikeResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.gamemate.domain.like.dto;
+package com.example.gamemate.domain.like.dto.response;
 
 import com.example.gamemate.domain.like.entity.ReviewLike;
 import lombok.Getter;
