+++ src/main/java/com/example/gamemate/domain/review/dto/request/ReviewCreateRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.gamemate.domain.review.dto;
+package com.example.gamemate.domain.review.dto.request;
 
 import lombok.Getter;
