+++ src/main/resources/application.properties
@@ -71,3 +71,13 @@ logging.level.com.example.gamemate=DEBUG
 gemini.api.url=${GEMINI_URL}
 gemini.api.key=${GEMINI_KEY}
 
+#S3
+cloud.aws.credentials.access-key=${AWS_ACCESS_KEY}
+cloud.aws.credentials.secret-key=${AWS_SECRET_KEY}
+cloud.aws.s3.bucket=${AWS_BUCKET}
+cloud.aws.region.static=${AWS_REGION}
+cloud.aws.stack.auto=${AWS_STACK_AUTO}
+
+#multipart
+spring.servlet.multipart.max-file-size=5MB
+spring.servlet.multipart.max-request-size=5MB
